Put big buffers on the stack.
Let me tell you a story:  It all started when I was trying to get DrawHDMI
to work the first time back when I drew a purple square on a blue screen.

I noticed that the program initialization was not working but I started
from a working project and all I did was to add code that wasn't even
used.

One difference that could have an effect was that in the .c file that I
added, I declared a big buffer to hold the screen's data without removing
the one that was in cf_hdmi.c.  So I tried to remove the declaration of
the unused buffer.

This caused the program to start working again.  Therefore, I deduced that
having those big buffers changes how the memory of the initial porgram is
layed out.

I put the big buffers on the stack.  How do I do this? I have to declare
them in a function from which we never return.  This is why I HAVE to
declare them in the main() function.

As you can see, in the main, I declare three big buffers on the stack and
pass hand them over to the modules that will use them.

NOTES:

The preceding commit works, that is because I reduced the size of the
buffers that were going to hold the data from the bitmaps for the pieces
and for the characters from 800 000 and 200 000 to 8 and 2.

From that commit, changing only the size of these buffers causes the
screen to display crap instead of the chessboard that it should display.

BEFORE I TRIED THIS:

Before I tried this, I tried playing around with the linker script.  The
stack size was initially set to 0x2000 and I set it to 0x0fff ffff which
made the screen show something more coherent.  However there was crap in a
good part of the screen.

TLDR: The preceding commit was tested but setting the buffer sizes back to
their original values of 800k and 200k will make the screen be filled with
garbage.

This commit explores the solution of putting the buffers on the stack in
order for the memory of the program to be smaller.

Better understanding of the linker script would be nice but the solution
of putting the big buffers on the stack is OK even though it puts silly
looking stuff in the main.