Drawing bitmaps in the screen.
the index conversion for bitmaps is slightly weird.  Only a small mistake to
correct.

I'm going to try to figure out what the problem is.

Note that the (0,0) of the bitmap is at the bottom left, that's why we put
data[(bmp->Height - b_i - 1) * bmp->Width  + b_j ) at pixel (i,j) in the screen.