The board display is now super awesome!
To test, the reqest

   ${IP_address}/status/board

can be used, each time it is called, it will make a BoardDisplay_move_piece()
call to test all the stuff that the boardDisplay module offers.

Also, the test_move_piece() function gives an example of how we transfer
information to the boardDisplay module in order to make a move.