Upgraded BoardDisplay to take into account castling and en-passant capture.
Note: caller will have to specify if the move is a capturing move
for the notation to be written appropriately.  This should not be
a problem since the chessboard logic has to figure this out before
calling BoardDisplay_move_piece.

Also, it will have to specify if the move is an en-passant capture.