Standardized BoardDisplay interface.
All interface functions will call cf_hdmi_send_buffer() so that the client does
not have to worry about that since it indeed should not have to worry about
that.

Note the change in chessboard.c: in start_game, we call
BoardDisplay_start_game() which will add info to show that the game that was
created previously is now started (It will write "white to move" to the screen).

Also, now BoardDisplay_move_piece draws "White to move" or "Black to move"
automatically.  (Yay encapsulation).

Moved functions around in the source files so that they appear in a more
coherent order.

Added the static keyword to the functions that are to be private.