A move is refused if the current player's king is in check after the proposed move
In chessboard.c, in the movePiece function, the move, if it is legal with
regards to other requirements, the move is "tried" then we look if the king is
in check.  If that is the case, then the move is undone and the function returns
ILLEGAL without changing whose turn it is, or changing the move number etc...

No calls to the display module are made either.

From this commit, it can be tested using the push button.
1st press, start game, make a couple moves
2nd press, move the bishop to put the king in check
3rd press, try a move that doesn't resolve the check (move is refuset)
4th press, the move does resolve the check and is accepted.