Four big changes here.
First, the enPassant feature is almost complete and works.
What is missing is the HDMI screen update.

Second, the promotion is changed so it works more like a revive mechanic:
you can only promote to an already captured allied piece and, internally,
we use the new type's piece space. The corresponding test was updated, too.

Third, the JSON parser was changed so that it can accept both 'yes|no' and
'true|false' logic, since both can be found.

Fourth, a 'REDUCE_SPAM' macro was added in the server_tcp file which,
when defined, reduce the amount of printed request to a minimum. More
precisely, /status/board, /status/summary and /game_details are not
displayed.